Fix add teacher button and split widget
