+++ main_window.py
@@ -1,12 +1,7 @@
-from PyQt6.QtWidgets import QMainWindow, QWidget
+from PyQt6.QtWidgets import QMainWindow
 from PyQt6 import uic
 
-
-class TeacherItem(QWidget):
-    def __init__(self, index=1, parent=None):
-        super().__init__(parent)
-        uic.loadUi("teachersItem.ui", self)
-        self.index_label.setText(str(index))
+from teacher_item import TeacherItem
 
 class MainWindow(QMainWindow):
     def __init__(self):
@@ -17,9 +12,15 @@ class MainWindow(QMainWindow):
         self.stackedWidget.addWidget(self.teachers_page)
         self.teachers_page.teachersListLayout.addWidget(TeacherItem())
 
+        self.teachers_page.add_btn.clicked.connect(self._add_teacher)
+
         self.teachers_btn_1.toggled.connect(self._show_teachers)
         self.teachers_btn_2.toggled.connect(self._show_teachers)
 
     def _show_teachers(self, checked):
         if checked:
             self.stackedWidget.setCurrentWidget(self.teachers_page)
+
+    def _add_teacher(self):
+        index = self.teachers_page.teachersListLayout.count() + 1
+        self.teachers_page.teachersListLayout.addWidget(TeacherItem(index))
